Simplify HTML5 load method drastically
